Moved integration Tests, added placeholder for  Drug Controller integration test
